Refresh Toekn and Access Token completed

+++ src/utils/cloudinary.js
@@ -14,16 +14,20 @@ const uploadOnCloudinary = async (localFilePath) => {
     try {
         if(!localFilePath) return null
         //upload file in cloudinary
-        const responce = awaitcloudinary.uploader.upload(localFilePath, {
-            resource_type : auto  //Which type of data you want to upload like pdf, image or vdo etc auto for everything allowed
+        const responce = await cloudinary.uploader.upload(localFilePath, {
+            resource_type : "auto"  //Which type of data you want to upload like pdf, image or vdo etc auto for everything allowed
         })
         //file has been uploaded successfully
-        console.log("File is uploaded on cloudinary", responce.url)
+        //console.log("File is uploaded on cloudinary", responce.url)
+        fs.unlinkSync(localFilePath);
+        console.log("Responce : ",responce);
         return responce;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) //remove the locally saved temporary files as the upload operation got failed
-        return null;
+        if(fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath) //remove the locally saved temporary files as the upload operation got failed
+            return null;
+        }
     }
 }
